# Cache refactoring for performance improvements