--- a/composer.json
+++ b/composer.json
@@ -31,10 +31,10 @@
         "php-di/phpdoc-reader": "^2.0.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "~5.0",
-        "mnapoli/phpunit-easymock": "~0.2.0",
-        "doctrine/annotations": "~1.2",
-        "ocramius/proxy-manager": "~2.0.2"
+        "phpunit/phpunit": "^5.7.0",
+        "mnapoli/phpunit-easymock": "^0.2.3",
+        "doctrine/annotations": "^1.2.0",
+        "ocramius/proxy-manager": "^2.0.2"
     },
     "replace": {
         "mnapoli/php-di": "*"