--- a/tests/UnitTest/Definition/Source/CachedDefinitionSourceTest.php
+++ b/tests/UnitTest/Definition/Source/CachedDefinitionSourceTest.php
@@ -5,8 +5,8 @@ namespace DI\Test\UnitTest\Definition\Source;
 use DI\Definition\ObjectDefinition;
 use DI\Definition\Source\CachedDefinitionSource;
 use DI\Definition\Source\DefinitionArray;
+use DI\Definition\Source\DefinitionSource;
 use EasyMock\EasyMock;
-use Psr\SimpleCache\CacheInterface;
 
 /**
  * @covers \DI\Definition\Source\CachedDefinitionSource
@@ -15,50 +15,32 @@ class CachedDefinitionSourceTest extends \PHPUnit_Framework_TestCase
 {
     use EasyMock;
 
+    public function setUp()
+    {
+        parent::setUp();
+
+        if (!CachedDefinitionSource::isSupported()) {
+            $this->markTestSkipped('APCu extension is required');
+        }
+
+        apcu_clear_cache();
+    }
+
     /**
      * @test
      */
     public function should_get_from_cache()
     {
-        /** @var CacheInterface $cache */
-        $cache = $this->easySpy(CacheInterface::class, [
-            'get' => 'foo',
-        ]);
+        $source = $this->createMock(DefinitionSource::class);
+        $source
+            ->expects($this->once()) // The sub-source should be called ONLY ONCE
+            ->method('getDefinition')
+            ->willReturn('bar');
 
-        $source = new CachedDefinitionSource(new DefinitionArray(), $cache);
+        $source = new CachedDefinitionSource($source);
 
-        $this->assertEquals('foo', $source->getDefinition('foo'));
-    }
-
-    /**
-     * @dataProvider cacheKeyProvider
-     */
-    public function testCacheKey($rawName, $cacheName)
-    {
-        $cache = $this->easySpy(CacheInterface::class);
-
-        $source = new CachedDefinitionSource(new DefinitionArray(), $cache);
-
-        $cache->expects($this->once())
-            ->method('get')
-            ->with(CachedDefinitionSource::CACHE_PREFIX . $cacheName)
-            ->will($this->returnValue(false));
-
-        $cache->expects($this->once())
-            ->method('set')
-            ->with(CachedDefinitionSource::CACHE_PREFIX . $cacheName);
-
-        $source->getDefinition($rawName);
-    }
-
-    public function cacheKeyProvider()
-    {
-        return [
-            ['foo{bar}baz', 'foo.bar.baz'],
-            ['foo(bar)baz', 'foo.bar.baz'],
-            ['foo/bar\baz', 'foo.bar.baz'],
-            ['foo@bar:baz', 'foo.bar.baz'],
-        ];
+        self::assertEquals('bar', $source->getDefinition('foo'));
+        self::assertEquals('bar', $source->getDefinition('foo'));
     }
 
     /**
@@ -66,22 +48,20 @@ class CachedDefinitionSourceTest extends \PHPUnit_Framework_TestCase
      */
     public function should_save_to_cache_and_return()
     {
-        $cache = $this->easySpy(CacheInterface::class, [
-            'get' => false,
-        ]);
-
         $cachedSource = new DefinitionArray([
             'foo' => \DI\create(),
         ]);
 
-        $source = new CachedDefinitionSource($cachedSource, $cache);
+        $source = new CachedDefinitionSource($cachedSource);
 
-        $expectedDefinition = new ObjectDefinition('foo');
-        $cache->expects($this->once())
-            ->method('set')
-            ->with($this->isType('string'), $expectedDefinition);
+        // Sanity check
+        self::assertSavedInCache('foo', null);
 
-        $this->assertEquals($expectedDefinition, $source->getDefinition('foo'));
+        // Return the definition
+        self::assertEquals(new ObjectDefinition('foo'), $source->getDefinition('foo'));
+
+        // The definition is saved in the cache
+        self::assertSavedInCache('foo', new ObjectDefinition('foo'));
     }
 
     /**
@@ -89,15 +69,15 @@ class CachedDefinitionSourceTest extends \PHPUnit_Framework_TestCase
      */
     public function should_save_null_to_cache_and_return_null()
     {
-        $cache = $this->easySpy(CacheInterface::class, [
-            'get' => false,
-        ]);
+        $source = new CachedDefinitionSource(new DefinitionArray());
 
-        $source = new CachedDefinitionSource(new DefinitionArray(), $cache);
+        self::assertNull($source->getDefinition('foo'));
+        self::assertSavedInCache('foo', null);
+    }
 
-        $cache->expects($this->once())
-            ->method('set')
-            ->with($this->isType('string'), null);
-        $this->assertNull($source->getDefinition('foo'));
+    private static function assertSavedInCache(string $definitionName, $expectedValue)
+    {
+        $definitions = apcu_fetch(CachedDefinitionSource::CACHE_KEY);
+        self::assertEquals($expectedValue, $definitions[$definitionName]);
     }
 }