--- a/travis.php.ini
+++ b/travis.php.ini
@@ -1,0 +1,3 @@
+extension="apcu.so"
+# Enable APC on the CLI to run all tests
+apc.enable_cli=1