🎉   yuuuuge performance improvements incoming :)

Fixes #477 This pull request refactors the whole cache system.

## Changes

Before:

- every (cacheable) definition is cached as a separate entry
- each definition is `get` on demand
- each definition is `set` when not found
- all PSR-16 caches are supported
- APCu is recommended because given the number of get/set to the cache, all other caches are too slow

After:

- only APCu is supported: that allows a much more optimized code, less errors from users, and a more "plug and play" solution (just enable the cache, nothing to configure)
- all definitions are cached under a single cache entry
- all the definitions are `get` at once: huge gain
- each definition is `set` when not found (not optimisation here)

Things to look out for: cache stampede.

## Performance comparison

### Benchmark `get-cache.php`

Before:

- no cache used: 224ms
- APCu cache used: 181ms
- gain thanks to the cache: **19%**
- [profiling](https://blackfire.io/profiles/compare/702a492a-f462-4aed-b210-8ab7ecf2d1eb/graph)

After:

- no cache used: 225ms
- APCu cache used: 146ms
- gain thanks to the cache: **35%**
- [profiling](https://blackfire.io/profiles/compare/29367b6a-b2c5-4e4b-8d98-aec20a1c8feb/graph)

When you use the cache there is a **19% improvement** (181 -> 146ms). [Profiling](https://blackfire.io/profiles/compare/cf181d0a-7629-4d21-aee6-ed78550573f4/graph)

## TODO

- [x] POC
- [x] ~~optimize APCu calls to avoid many `set` in the same `Container::get()` call~~ Not interesting
- [ ] test with 3rd party DI container benchmarks
- [ ] update the `ContainerBuilder`
- [ ] configurable cache key prefix
- [ ] the cache key should contain a hash of all sources? auto-refresh?
- [ ] decide what to do with `ArrayCache`
- [ ] update documentation
- [ ] changelog