@mnapoli 2017-05-24
That is not really reassuring though: https://twitter.com/Ocramius/status/867477833123328000

> Yes, but then I would never use APC/APCu for anything under high write load for the same reason

https://twitter.com/Ocramius/status/867477517623599105

> Usually just the write being fast. APCu often gets cache corruption when under heavy write load, so we're under the same league