@mnapoli 2017-06-04
Made obsolete by #494